Create a simple TestPolicy to sub-class from when testing Policies (reduces code in some test cases).
